Refactor level set module, remove magic numbers, improve documentation
