+++ CHECKLIST.md
@@ -1,6 +1,6 @@
 # CFD Suite - Technical Checklist
 
-## Version 0.60.0 - Current State
+## Version 0.61.0 - Current State
 
 ### Completed ✅
 - [x] Workspace builds without errors
@@ -30,6 +30,16 @@
 - [x] All public APIs fully documented
 - [x] Dead code eliminated and all functions properly exposed
 - [x] Build warnings resolved
+
+### Completed (v0.61.0) ✅
+- [x] Deep architectural review completed
+- [x] Removed duplicate numerical_validation.rs file
+- [x] Refactored level_set module into proper modular structure
+- [x] Fixed remaining magic numbers (TWO, THREE, FOUR, etc.)
+- [x] Added documentation for all enum variants
+- [x] No stubs, unimplemented!, todo!, or panic! found
+- [x] Validated all physics against literature references
+- [x] All modules now <500 LOC with proper separation
 
 ### Completed (v0.60.0) ✅
 - [x] Fixed naming violations (temp_fields → state_buffer, f_temp → f_buffer)
