+++ PRD.md
@@ -1,6 +1,6 @@
 # Product Requirements Document
 
-## CFD Suite v0.60.0
+## CFD Suite v0.61.0
 
 ### Product Classification
 Research software (not production)
