+++ crates/cfd-3d/src/level_set/solver.rs
@@ -1,0 +1,117 @@
+//! Level Set solver implementation
+
+use super::config::LevelSetConfig;
+use cfd_core::error::Result;
+use nalgebra::{RealField, Vector3};
+use num_traits::FromPrimitive;
+
+/// Level Set solver for interface tracking
+pub struct LevelSetSolver<T: RealField + FromPrimitive + Copy> {
+    config: LevelSetConfig,
+    /// Grid dimensions
+    nx: usize,
+    ny: usize,
+    nz: usize,
+    /// Grid spacing
+    dx: T,
+    dy: T,
+    dz: T,
+    /// Level set function (signed distance)
+    phi: Vec<T>,
+    /// Previous timestep level set function
+    phi_previous: Vec<T>,
+    /// Velocity field
+    velocity: Vec<Vector3<T>>,
+    /// Narrow band indices (if using narrow band)
+    narrow_band: Vec<usize>,
+    /// Time step counter
+    time_step: usize,
+}
+
+impl<T: RealField + FromPrimitive + Copy> LevelSetSolver<T> {
+    /// Create a new Level Set solver
+    pub fn new(
+        config: LevelSetConfig,
+        nx: usize,
+        ny: usize,
+        nz: usize,
+        dx: T,
+        dy: T,
+        dz: T,
+    ) -> Self {
+        let grid_size = nx * ny * nz;
+        Self {
+            config,
+            nx,
+            ny,
+            nz,
+            dx,
+            dy,
+            dz,
+            phi: vec![T::zero(); grid_size],
+            phi_previous: vec![T::zero(); grid_size],
+            velocity: vec![Vector3::zeros(); grid_size],
+            narrow_band: Vec::new(),
+            time_step: 0,
+        }
+    }
+
+    /// Get grid index from (i, j, k) coordinates
+    #[inline]
+    pub fn index(&self, i: usize, j: usize, k: usize) -> usize {
+        k * self.ny * self.nx + j * self.nx + i
+    }
+
+    /// Get the level set function
+    pub fn phi(&self) -> &[T] {
+        &self.phi
+    }
+
+    /// Get mutable reference to level set function
+    pub fn phi_mut(&mut self) -> &mut [T] {
+        &mut self.phi
+    }
+
+    /// Set velocity field
+    pub fn set_velocity(&mut self, velocity: Vec<Vector3<T>>) {
+        self.velocity = velocity;
+    }
+
+    /// Get narrow band indices
+    pub fn narrow_band(&self) -> &[usize] {
+        &self.narrow_band
+    }
+
+    /// Update narrow band indices based on current level set
+    pub fn update_narrow_band(&mut self) {
+        self.narrow_band.clear();
+        let band_width = T::from_f64(self.config.band_width).unwrap_or_else(T::zero);
+
+        for idx in 0..self.phi.len() {
+            if self.phi[idx].abs() <= band_width * self.dx.min(self.dy).min(self.dz) {
+                self.narrow_band.push(idx);
+            }
+        }
+    }
+
+    /// Advance level set by one time step
+    pub fn advance(&mut self, dt: T) -> Result<()> {
+        // Store previous level set
+        self.phi_previous.copy_from_slice(&self.phi);
+
+        // Advection step would be implemented here
+        // This is a placeholder for the actual advection implementation
+
+        // Check for reinitialization
+        self.time_step += 1;
+        if self.time_step % self.config.reinitialization_interval == 0 {
+            // Reinitialization would be called here
+        }
+
+        if self.config.use_narrow_band {
+            self.update_narrow_band();
+        }
+
+        Ok(())
+    }
+}
