+++ crates/cfd-validation/src/lib.rs
@@ -15,7 +15,6 @@ pub mod convergence;
 pub mod error_metrics;
 pub mod literature;
 pub mod numerical;
-pub mod numerical_validation; // Kept for backward compatibility, will be deprecated
 pub mod solutions;
 pub mod time_integration_validation;
