--- crates/cfd-validation/src/numerical_validation.rs
@@ -1,721 +1,0 @@
-//! Numerical algorithm validation against literature solutions.
-//!
-//! This module provides comprehensive validation tests for numerical algorithms
-//! implemented in the CFD suite, comparing against known analytical solutions
-//! and published benchmark results.
-
-use cfd_core::error::{Error, Result};
-use cfd_math::linear_solver::LinearSolverConfig;
-use cfd_math::{BiCGSTAB, ConjugateGradient, LinearSolver};
-use nalgebra::{DVector, RealField};
-use nalgebra_sparse::CsrMatrix;
-use num_traits::{Float, FromPrimitive};
-
-/// Validation result for a numerical algorithm
-#[derive(Debug, Clone)]
-pub struct ValidationResult<T: RealField + Copy> {
-    /// Algorithm name
-    pub algorithm_name: String,
-    /// Test case name
-    pub test_case: String,
-    /// Computed solution
-    pub computed_solution: DVector<T>,
-    /// Analytical solution
-    pub analytical_solution: DVector<T>,
-    /// Error metrics
-    pub error_metrics: ErrorMetrics<T>,
-    /// Convergence information
-    pub convergence_info: ConvergenceInfo<T>,
-    /// Literature reference
-    pub literature_reference: String,
-    /// Test passed
-    pub passed: bool,
-}
-
-/// Error metrics for validation
-#[derive(Debug, Clone)]
-pub struct ErrorMetrics<T: RealField + Copy> {
-    /// L2 norm of error
-    pub l2_error: T,
-    /// L∞ norm of error
-    pub linf_error: T,
-    /// Relative L2 error
-    pub relative_l2_error: T,
-    /// Root mean square error
-    pub rmse: T,
-}
-
-/// Convergence information
-#[derive(Debug, Clone)]
-pub struct ConvergenceInfo<T: RealField + Copy> {
-    /// Number of iterations
-    pub iterations: usize,
-    /// Final residual
-    pub final_residual: T,
-    /// Convergence rate
-    pub convergence_rate: Option<T>,
-}
-
-/// Linear solver validation suite
-pub struct LinearSolverValidator;
-
-impl LinearSolverValidator {
-    /// Validate linear solvers against analytical solutions
-    pub fn validate_all<T: RealField + Copy + FromPrimitive + Copy + Float>(
-    ) -> Result<Vec<ValidationResult<T>>> {
-        let mut results = Vec::new();
-
-        // Test 1: Standard diagonal system
-        match Self::test_diagonal_system::<T>() {
-            Ok(test_results) => results.extend(test_results),
-            Err(e) => println!("Diagonal system test failed: {e}"),
-        }
-
-        // Test 2: Tridiagonal system (1D Poisson)
-        match Self::test_tridiagonal_system::<T>() {
-            Ok(test_results) => results.extend(test_results),
-            Err(e) => println!("Tridiagonal system test failed: {e}"),
-        }
-
-        // Test 3: 2D Poisson equation
-        match Self::test_2d_poisson::<T>() {
-            Ok(test_results) => results.extend(test_results),
-            Err(e) => println!("2D Poisson test failed: {e}"),
-        }
-
-        // Test 4: Ill-conditioned system (expected to have some failures)
-        match Self::test_ill_conditioned_system::<T>() {
-            Ok(test_results) => results.extend(test_results),
-            Err(e) => println!("Ill-conditioned system test failed (expected): {e}"),
-        }
-
-        Ok(results)
-    }
-
-    /// Test diagonal system: Ax = b where A is diagonal
-    /// Literature: Golub & Van Loan (2013), "Matrix Computations", 4th Edition
-    fn test_diagonal_system<T: RealField + Copy + FromPrimitive + Copy + Float>(
-    ) -> Result<Vec<ValidationResult<T>>> {
-        let n = 100;
-        let mut results = Vec::new();
-
-        // Create diagonal matrix A and RHS b
-        let (a, b, analytical) = Self::create_diagonal_system::<T>(n)?;
-
-        // Test different solvers
-        let solvers: Vec<(&str, Box<dyn LinearSolver<T>>)> = vec![
-            (
-                "ConjugateGradient",
-                Box::new(ConjugateGradient::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-        ];
-
-        for (name, solver) in solvers {
-            match solver.solve(&a, &b, None) {
-                Ok(computed) => {
-                    let error_metrics = Self::compute_error_metrics(&computed, &analytical);
-
-                    let result = ValidationResult {
-                        algorithm_name: name.to_string(),
-                        test_case: "Diagonal System".to_string(),
-                        computed_solution: computed,
-                        analytical_solution: analytical.clone(),
-                        error_metrics: error_metrics.clone(),
-                        convergence_info: ConvergenceInfo {
-                            iterations: 1, // Diagonal systems converge in 1 iteration
-                            final_residual: error_metrics.l2_error,
-                            convergence_rate: None,
-                        },
-                        literature_reference:
-                            "Golub & Van Loan (2013), Matrix Computations, 4th Ed.".to_string(),
-                        passed: error_metrics.relative_l2_error
-                            < T::from_f64(1e-12).unwrap_or_else(|| T::zero()),
-                    };
-                    results.push(result);
-                }
-                Err(e) => {
-                    println!("Solver {name} failed on diagonal system: {e}");
-
-                    // Return error report for solver failure
-                    let result = ValidationResult {
-                        algorithm_name: name.to_string(),
-                        test_case: "Diagonal System".to_string(),
-                        passed: false,
-                        computed_solution: DVector::from_element(
-                            analytical.len(),
-                            T::from_f64(f64::NAN).unwrap_or_else(T::zero),
-                        ),
-                        analytical_solution: analytical.clone(),
-                        error_metrics: ErrorMetrics {
-                            l2_error: T::from_f64(f64::INFINITY)
-                                .unwrap_or_else(|| T::from_f64(1e10).unwrap_or_else(T::one)),
-                            linf_error: T::from_f64(f64::INFINITY)
-                                .unwrap_or_else(|| T::from_f64(1e10).unwrap_or_else(T::one)),
-                            relative_l2_error: T::from_f64(f64::INFINITY)
-                                .unwrap_or_else(|| T::from_f64(1e10).unwrap_or_else(T::one)),
-                            rmse: T::from_f64(f64::INFINITY)
-                                .unwrap_or_else(|| T::from_f64(1e10).unwrap_or_else(T::one)),
-                        },
-                        convergence_info: ConvergenceInfo {
-                            iterations: 0,
-                            final_residual: T::from_f64(f64::INFINITY)
-                                .unwrap_or_else(|| T::from_f64(1e10).unwrap_or_else(T::one)),
-                            convergence_rate: Some(T::zero()),
-                        },
-                        literature_reference:
-                            "Golub & Van Loan (2013), Matrix Computations, 4th Ed.".to_string(),
-                    };
-                    results.push(result);
-                }
-            }
-        }
-
-        Ok(results)
-    }
-
-    /// Test tridiagonal system (1D Poisson equation)
-    /// Literature: Strang (2007), "Computational Science and Engineering"
-    fn test_tridiagonal_system<T: RealField + Copy + FromPrimitive + Copy + Float>(
-    ) -> Result<Vec<ValidationResult<T>>> {
-        let n = 64;
-        let mut results = Vec::new();
-
-        // Create 1D Poisson system: -u'' = f with u(0) = u(1) = 0
-        let (a, b, analytical) = Self::create_1d_poisson_system::<T>(n)?;
-
-        let solvers: Vec<(&str, Box<dyn LinearSolver<T>>)> = vec![
-            (
-                "ConjugateGradient",
-                Box::new(ConjugateGradient::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-        ];
-
-        for (name, solver) in solvers {
-            let computed = solver.solve(&a, &b, None)?;
-            let error_metrics = Self::compute_error_metrics(&computed, &analytical);
-
-            let result = ValidationResult {
-                algorithm_name: name.to_string(),
-                test_case: "1D Poisson Equation".to_string(),
-                computed_solution: computed,
-                analytical_solution: analytical.clone(),
-                error_metrics: error_metrics.clone(),
-                convergence_info: ConvergenceInfo {
-                    iterations: 50, // Typical for CG on Poisson
-                    final_residual: error_metrics.l2_error,
-                    convergence_rate: Some(T::from_f64(0.95).unwrap_or_else(|| T::zero())), // Typical for CG
-                },
-                literature_reference: "Strang (2007), Computational Science and Engineering"
-                    .to_string(),
-                passed: error_metrics.relative_l2_error
-                    < T::from_f64(1e-10).unwrap_or_else(|| T::zero()),
-            };
-            results.push(result);
-        }
-
-        Ok(results)
-    }
-
-    /// Test 2D Poisson equation
-    /// Literature: `LeVeque` (2007), "Finite Difference Methods for ODEs and PDEs"
-    fn test_2d_poisson<T: RealField + Copy + FromPrimitive + Copy + Float>(
-    ) -> Result<Vec<ValidationResult<T>>> {
-        let nx = 32;
-        let ny = 32;
-        let mut results = Vec::new();
-
-        // Create 2D Poisson system with manufactured solution
-        let (a, b, analytical) = Self::create_2d_poisson_system::<T>(nx, ny)?;
-
-        let solvers: Vec<(&str, Box<dyn LinearSolver<T>>)> = vec![
-            (
-                "ConjugateGradient",
-                Box::new(ConjugateGradient::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-        ];
-
-        for (name, solver) in solvers {
-            let computed = solver.solve(&a, &b, None)?;
-            let error_metrics = Self::compute_error_metrics(&computed, &analytical);
-
-            let result = ValidationResult {
-                algorithm_name: name.to_string(),
-                test_case: "2D Poisson Equation".to_string(),
-                computed_solution: computed,
-                analytical_solution: analytical.clone(),
-                error_metrics: error_metrics.clone(),
-                convergence_info: ConvergenceInfo {
-                    iterations: 100, // Typical for 2D Poisson
-                    final_residual: error_metrics.l2_error,
-                    convergence_rate: Some(T::from_f64(0.98).unwrap_or_else(|| T::zero())),
-                },
-                literature_reference: "LeVeque (2007), Finite Difference Methods for ODEs and PDEs"
-                    .to_string(),
-                passed: error_metrics.relative_l2_error
-                    < T::from_f64(1e-8).unwrap_or_else(|| T::zero()),
-            };
-            results.push(result);
-        }
-
-        Ok(results)
-    }
-
-    /// Test ill-conditioned system
-    /// Literature: Higham (2002), "Accuracy and Stability of Numerical Algorithms"
-    fn test_ill_conditioned_system<T: RealField + Copy + FromPrimitive + Copy + Float>(
-    ) -> Result<Vec<ValidationResult<T>>> {
-        let n = 50;
-        let mut results = Vec::new();
-
-        // Create Hilbert matrix (ill-conditioned)
-        let (a, b, analytical) = Self::create_hilbert_system::<T>(n)?;
-
-        // Only test robust solvers for ill-conditioned systems
-        let solvers: Vec<(&str, Box<dyn LinearSolver<T>>)> = vec![
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-            (
-                "BiCGSTAB",
-                Box::new(BiCGSTAB::new(LinearSolverConfig::default())),
-            ),
-        ];
-
-        for (name, solver) in solvers {
-            // Handle potential solver breakdown gracefully
-            match solver.solve(&a, &b, None) {
-                Ok(computed) => {
-                    let error_metrics = Self::compute_error_metrics(&computed, &analytical);
-
-                    let result = ValidationResult {
-                        algorithm_name: name.to_string(),
-                        test_case: "Ill-Conditioned System (Hilbert)".to_string(),
-                        computed_solution: computed,
-                        analytical_solution: analytical.clone(),
-                        error_metrics: error_metrics.clone(),
-                        convergence_info: ConvergenceInfo {
-                            iterations: 200, // More iterations for ill-conditioned
-                            final_residual: error_metrics.l2_error,
-                            convergence_rate: Some(T::from_f64(0.99).unwrap_or_else(|| T::zero())),
-                        },
-                        literature_reference:
-                            "Higham (2002), Accuracy and Stability of Numerical Algorithms"
-                                .to_string(),
-                        passed: error_metrics.relative_l2_error
-                            < T::from_f64(1e-6).unwrap_or_else(|| T::zero()), // Relaxed tolerance
-                    };
-                    results.push(result);
-                }
-                Err(e) => {
-                    // For ill-conditioned systems, solver breakdown is expected for some methods
-                    println!("Solver {name} failed on ill-conditioned system (expected): {e}");
-
-                    // Create a failed result entry
-                    let dummy_solution = DVector::zeros(analytical.len());
-                    let error_metrics = Self::compute_error_metrics(&dummy_solution, &analytical);
-
-                    let result = ValidationResult {
-                        algorithm_name: name.to_string(),
-                        test_case: "Ill-Conditioned System (Hilbert)".to_string(),
-                        computed_solution: dummy_solution,
-                        analytical_solution: analytical.clone(),
-                        error_metrics: error_metrics.clone(),
-                        convergence_info: ConvergenceInfo {
-                            iterations: 0,
-                            final_residual: T::from_f64(f64::INFINITY).unwrap_or_else(|| T::zero()),
-                            convergence_rate: None,
-                        },
-                        literature_reference:
-                            "Higham (2002), Accuracy and Stability of Numerical Algorithms"
-                                .to_string(),
-                        passed: false, // Mark as failed due to breakdown
-                    };
-                    results.push(result);
-                }
-            }
-        }
-
-        Ok(results)
-    }
-
-    /// Create diagonal system for testing
-    fn create_diagonal_system<T: RealField + Copy + FromPrimitive + Copy>(
-        n: usize,
-    ) -> Result<(CsrMatrix<T>, DVector<T>, DVector<T>)> {
-        // Create diagonal matrix with entries 1, 2, 3, ..., n using iterators
-        let diagonal_entries: Vec<(usize, usize, T)> = (0..n)
-            .map(|i| (i, i, T::from_usize(i + 1).unwrap_or_else(|| T::zero())))
-            .collect();
-
-        let (row_indices, col_indices, values): (Vec<_>, Vec<_>, Vec<_>) =
-            diagonal_entries.into_iter().fold(
-                (Vec::new(), Vec::new(), Vec::new()),
-                |(mut rows, mut cols, mut vals), (r, c, v)| {
-                    rows.push(r);
-                    cols.push(c);
-                    vals.push(v);
-                    (rows, cols, vals)
-                },
-            );
-
-        let a = {
-            // Convert triplets to CSR format using iterator combinators
-            let mut row_offsets = vec![0; n + 1];
-            let mut sorted_data: Vec<(usize, usize, T)> = row_indices
-                .into_iter()
-                .zip(col_indices)
-                .zip(values)
-                .map(|((r, c), v)| (r, c, v))
-                .collect();
-
-            sorted_data.sort_by_key(|(r, c, _)| (*r, *c));
-
-            let mut current_row = 0;
-            let mut csr_col_indices = Vec::new();
-            let mut csr_values = Vec::new();
-
-            for (r, c, v) in sorted_data {
-                while current_row < r {
-                    current_row += 1;
-                    row_offsets[current_row] = csr_col_indices.len();
-                }
-                csr_col_indices.push(c);
-                csr_values.push(v);
-            }
-
-            while current_row < n {
-                current_row += 1;
-                row_offsets[current_row] = csr_col_indices.len();
-            }
-
-            CsrMatrix::try_from_csr_data(n, n, row_offsets, csr_col_indices, csr_values)
-                .map_err(|_| Error::InvalidConfiguration("Matrix solve failed".into()))?
-        };
-
-        // Create RHS b = [1, 1, ..., 1]
-        let b = DVector::from_element(n, T::one());
-
-        // Analytical solution: x[i] = 1 / (i + 1)
-        let analytical = DVector::from_iterator(
-            n,
-            (0..n).map(|i| T::one() / T::from_usize(i + 1).unwrap_or_else(|| T::zero())),
-        );
-
-        Ok((a, b, analytical))
-    }
-
-    /// Create 1D Poisson system
-    fn create_1d_poisson_system<T: RealField + Copy + FromPrimitive + Copy>(
-        n: usize,
-    ) -> Result<(CsrMatrix<T>, DVector<T>, DVector<T>)> {
-        let h = T::one() / T::from_usize(n + 1).unwrap_or_else(|| T::zero());
-        let h_squared = h * h;
-
-        // Create tridiagonal matrix for -u'' using iterators
-        let diagonal_value = T::from_f64(2.0).unwrap_or_else(|| T::zero()) / h_squared;
-        let off_diagonal_value = -T::one() / h_squared;
-
-        let (row_indices, col_indices, values): (Vec<_>, Vec<_>, Vec<_>) = (0..n)
-            .flat_map(|i| {
-                let mut entries = vec![(i, i, diagonal_value)];
-
-                if i > 0 {
-                    entries.push((i, i - 1, off_diagonal_value));
-                }
-                if i < n - 1 {
-                    entries.push((i, i + 1, off_diagonal_value));
-                }
-
-                entries
-            })
-            .fold(
-                (Vec::new(), Vec::new(), Vec::new()),
-                |(mut rows, mut cols, mut vals), (r, c, v)| {
-                    rows.push(r);
-                    cols.push(c);
-                    vals.push(v);
-                    (rows, cols, vals)
-                },
-            );
-
-        let a = {
-            // Convert triplets to CSR format
-            let mut row_offsets = vec![0; n + 1];
-            let mut sorted_data: Vec<(usize, usize, T)> = row_indices
-                .into_iter()
-                .zip(col_indices)
-                .zip(values)
-                .map(|((r, c), v)| (r, c, v))
-                .collect();
-
-            sorted_data.sort_by_key(|(r, c, _)| (*r, *c));
-
-            let mut current_row = 0;
-            let mut csr_col_indices = Vec::new();
-            let mut csr_values = Vec::new();
-
-            for (r, c, v) in sorted_data {
-                while current_row < r {
-                    current_row += 1;
-                    row_offsets[current_row] = csr_col_indices.len();
-                }
-                csr_col_indices.push(c);
-                csr_values.push(v);
-            }
-
-            while current_row < n {
-                current_row += 1;
-                row_offsets[current_row] = csr_col_indices.len();
-            }
-
-            CsrMatrix::try_from_csr_data(n, n, row_offsets, csr_col_indices, csr_values)
-                .map_err(|_| Error::InvalidConfiguration("Matrix solve failed".into()))?
-        };
-
-        // RHS for manufactured solution u(x) = x(1-x)
-        let b = DVector::from_iterator(
-            n,
-            (1..=n).map(|_i| {
-                // For the manufactured solution u(x) = x(1-x), the Laplacian is -2
-                T::from_f64(2.0).unwrap_or_else(|| T::zero())
-            }),
-        );
-
-        // Analytical solution
-        let analytical = DVector::from_iterator(
-            n,
-            (1..=n).map(|i| {
-                let x = T::from_usize(i).unwrap_or_else(|| T::zero()) * h;
-                x * (T::one() - x)
-            }),
-        );
-
-        Ok((a, b, analytical))
-    }
-
-    /// Create 2D Poisson system with 5-point stencil discretization
-    /// Solves: -∇²u = f on unit square with Dirichlet boundary conditions
-    fn create_2d_poisson_system<T: RealField + Copy + FromPrimitive + Copy>(
-        nx: usize,
-        ny: usize,
-    ) -> Result<(CsrMatrix<T>, DVector<T>, DVector<T>)> {
-        let n = nx * ny;
-        let h = T::one() / T::from_usize(nx - 1).unwrap_or_else(|| T::zero());
-        let h2 = h * h;
-
-        // Build sparse matrix using 5-point stencil
-        let mut row_offsets = vec![0];
-        let mut col_indices = Vec::new();
-        let mut values = Vec::new();
-
-        // Process each grid point
-        for idx in 0..n {
-            let i = idx % nx;
-            let j = idx / nx;
-
-            if i == 0 || i == nx - 1 || j == 0 || j == ny - 1 {
-                // Boundary point: u = 0 (identity row)
-                col_indices.push(idx);
-                values.push(T::one());
-            } else {
-                // Interior point: 5-point stencil
-                let center_coeff = T::from_f64(4.0).unwrap_or_else(|| T::zero()) / h2;
-                let neighbor_coeff = -T::one() / h2;
-
-                // Left neighbor
-                col_indices.push(idx - 1);
-                values.push(neighbor_coeff);
-
-                // Bottom neighbor
-                col_indices.push(idx - nx);
-                values.push(neighbor_coeff);
-
-                // Center
-                col_indices.push(idx);
-                values.push(center_coeff);
-
-                // Top neighbor
-                col_indices.push(idx + nx);
-                values.push(neighbor_coeff);
-
-                // Right neighbor
-                col_indices.push(idx + 1);
-                values.push(neighbor_coeff);
-            }
-
-            row_offsets.push(col_indices.len());
-        }
-
-        let a = CsrMatrix::try_from_csr_data(n, n, row_offsets, col_indices, values)
-            .map_err(|_| Error::InvalidConfiguration("Matrix construction failed".into()))?;
-
-        // Create RHS with manufactured solution u(x,y) = sin(πx)sin(πy)
-        let pi = T::from_f64(std::f64::consts::PI).unwrap_or_else(|| T::zero());
-        let mut b = DVector::zeros(n);
-        let mut analytical = DVector::zeros(n);
-
-        for idx in 0..n {
-            let i = idx % nx;
-            let j = idx / nx;
-            let x = T::from_usize(i).unwrap_or_else(|| T::zero()) * h;
-            let y = T::from_usize(j).unwrap_or_else(|| T::zero()) * h;
-
-            if i == 0 || i == nx - 1 || j == 0 || j == ny - 1 {
-                b[idx] = T::zero();
-                analytical[idx] = T::zero();
-            } else {
-                // f = 2π²sin(πx)sin(πy)
-                b[idx] = T::from_f64(2.0).unwrap_or_else(|| T::zero())
-                    * pi
-                    * pi
-                    * (pi * x).sin()
-                    * (pi * y).sin();
-                analytical[idx] = (pi * x).sin() * (pi * y).sin();
-            }
-        }
-
-        Ok((a, b, analytical))
-    }
-
-    /// Create Hilbert matrix system
-    fn create_hilbert_system<T: RealField + Copy + FromPrimitive + Copy>(
-        n: usize,
-    ) -> Result<(CsrMatrix<T>, DVector<T>, DVector<T>)> {
-        let mut row_indices = Vec::new();
-        let mut col_indices = Vec::new();
-        let mut values = Vec::new();
-
-        // Create Hilbert matrix H[i,j] = 1/(i+j+1)
-        for i in 0..n {
-            for j in 0..n {
-                row_indices.push(i);
-                col_indices.push(j);
-                values.push(T::one() / T::from_usize(i + j + 1).unwrap_or_else(|| T::zero()));
-            }
-        }
-
-        let a = {
-            // Convert triplets to CSR format
-            let mut row_offsets = vec![0; n + 1];
-            let mut sorted_data: Vec<(usize, usize, T)> = row_indices
-                .into_iter()
-                .zip(col_indices)
-                .zip(values)
-                .map(|((r, c), v)| (r, c, v))
-                .collect();
-
-            sorted_data.sort_by_key(|(r, c, _)| (*r, *c));
-
-            let mut current_row = 0;
-            let mut csr_col_indices = Vec::new();
-            let mut csr_values = Vec::new();
-
-            for (r, c, v) in sorted_data {
-                while current_row < r {
-                    current_row += 1;
-                    row_offsets[current_row] = csr_col_indices.len();
-                }
-                csr_col_indices.push(c);
-                csr_values.push(v);
-            }
-
-            while current_row < n {
-                current_row += 1;
-                row_offsets[current_row] = csr_col_indices.len();
-            }
-
-            CsrMatrix::try_from_csr_data(n, n, row_offsets, csr_col_indices, csr_values)
-                .map_err(|_| Error::InvalidConfiguration("Matrix solve failed".into()))?
-        };
-
-        // Use known solution x = [1, 1, ..., 1] and compute b = A*x
-        let x_true = DVector::from_element(n, T::one());
-        let b = &a * &x_true;
-
-        Ok((a, b, x_true))
-    }
-
-    /// Compute error metrics
-    fn compute_error_metrics<T: RealField + Copy + FromPrimitive + Copy>(
-        computed: &DVector<T>,
-        analytical: &DVector<T>,
-    ) -> ErrorMetrics<T> {
-        let diff = computed - analytical;
-
-        let l2_error = diff.norm();
-        let linf_error = diff.iter().map(|x| x.abs()).fold(T::zero(), T::max);
-
-        let analytical_norm = analytical.norm();
-        let relative_l2_error = if analytical_norm > T::zero() {
-            l2_error / analytical_norm
-        } else {
-            l2_error
-        };
-
-        let n = T::from_usize(computed.len()).unwrap_or_else(T::one);
-        let rmse = if n > T::zero() {
-            (l2_error * l2_error / n).sqrt()
-        } else {
-            l2_error
-        };
-
-        ErrorMetrics {
-            l2_error,
-            linf_error,
-            relative_l2_error,
-            rmse,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_linear_solver_validation() {
-        // For now, let's just test that the validation framework runs without crashing
-        // The actual solver issues need to be addressed separately
-        let results = LinearSolverValidator::validate_all::<f64>()
-            .expect("CRITICAL: Add proper error handling");
-
-        // Check that we have results (even if some failed)
-        assert!(!results.is_empty(), "Should have some validation results");
-
-        // Print results for debugging
-        for result in &results {
-            println!(
-                "Test: {} - {}: passed={}",
-                result.test_case, result.algorithm_name, result.passed
-            );
-        }
-
-        // For now, just ensure we have some results - the solver implementations
-        // may need refinement but the validation framework should work
-        println!(
-            "Linear solver validation completed with {} results",
-            results.len()
-        );
-    }
-}
